+++ banano-quest/banano-quest/Extensions/CGFloat+extension.swift
@@ -1,0 +1,16 @@
+//
+//  CGFloat+extension.swift
+//  banano-quest
+//
+//  Created by Luis De Leon on 8/3/18.
+//  Copyright © 2018 Michael O'Rourke. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension CGFloat {
+    static func random() -> CGFloat {
+        return CGFloat(arc4random()) / CGFloat(UInt32.max)
+    }
+}
