+++ banano-quest/banano-quest/Extensions/UIColor+extension.swift
@@ -1,0 +1,19 @@
+//
+//  UIColor+extension.swift
+//  banano-quest
+//
+//  Created by Luis De Leon on 8/3/18.
+//  Copyright © 2018 Michael O'Rourke. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UIColor {
+    static func random() -> UIColor {
+        return UIColor(red:   .random(),
+                       green: .random(),
+                       blue:  .random(),
+                       alpha: 1.0)
+    }
+}
