+++ banano-quest/banano-quest/ViewControllers/CreateQuestViewController.swift
@@ -13,21 +13,6 @@ import Pocket
 import MapKit
 import BigInt
 
-extension CGFloat {
-    static func random() -> CGFloat {
-        return CGFloat(arc4random()) / CGFloat(UInt32.max)
-    }
-}
-
-extension UIColor {
-    static func random() -> UIColor {
-        return UIColor(red:   .random(),
-                       green: .random(),
-                       blue:  .random(),
-                       alpha: 1.0)
-    }
-}
-
 class CreateQuestViewController: UIViewController, ColorPickerDelegate, UITextFieldDelegate {
     // UI Elements
     @IBOutlet weak var addLocationButton: UIButton!
